Update df-arcadedb to 23.10.1 (#24)

+++ server/src/test/java/com/arcadedb/server/ha/ReplicationServerReplicaRestartForceDbInstallIT.java
@@ -19,11 +19,13 @@
 package com.arcadedb.server.ha;
 
 import com.arcadedb.GlobalConfiguration;
+import com.arcadedb.database.Database;
+import com.arcadedb.engine.Bucket;
 import com.arcadedb.log.LogManager;
 import com.arcadedb.server.ArcadeDBServer;
-import com.arcadedb.server.TestCallback;
-import org.junit.jupiter.api.AfterEach;
+import com.arcadedb.server.ReplicationCallback;
 import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
 
 import java.io.*;
 import java.util.concurrent.atomic.*;
@@ -37,15 +39,7 @@ public class ReplicationServerReplicaRestartForceDbInstallIT extends Replication
   private          boolean    fullResync              = false;
 
   public ReplicationServerReplicaRestartForceDbInstallIT() {
-    GlobalConfiguration.HA_QUORUM.setValue("MAJORITY");
     GlobalConfiguration.HA_REPLICATION_QUEUE_SIZE.setValue(10);
-  }
-
-  @AfterEach
-  @Override
-  public void endTest() {
-    super.endTest();
-    GlobalConfiguration.HA_REPLICATION_QUEUE_SIZE.setValue(512);
   }
 
   @Override
@@ -57,15 +51,18 @@ public class ReplicationServerReplicaRestartForceDbInstallIT extends Replication
   @Override
   protected void onBeforeStarting(final ArcadeDBServer server) {
     if (server.getServerName().equals("ArcadeDB_2"))
-      server.registerTestEventListener(new TestCallback() {
+      server.registerTestEventListener(new ReplicationCallback() {
         @Override
         public void onEvent(final TYPE type, final Object object, final ArcadeDBServer server) {
+          if (!serversSynchronized)
+            return;
+
           if (slowDown) {
             // SLOW DOWN A SERVER AFTER 5TH MESSAGE
             if (totalMessages.incrementAndGet() > 5) {
               try {
-                LogManager.instance().log(this, Level.FINE, "TEST: Slowing down response from replica server 2...");
-                Thread.sleep(10000);
+                LogManager.instance().log(this, getErrorLevel(), "TEST: Slowing down response from replica server 2...");
+                Thread.sleep(10_000);
               } catch (final InterruptedException e) {
                 // IGNORE IT
                 LogManager.instance().log(this, Level.SEVERE, "TEST: ArcadeDB_2 HA event listener thread interrupted");
@@ -73,11 +70,12 @@ public class ReplicationServerReplicaRestartForceDbInstallIT extends Replication
               }
             }
           } else {
+
             if (type == TYPE.REPLICA_HOT_RESYNC) {
-              LogManager.instance().log(this, Level.FINE, "TEST: Received hot resync request");
+              LogManager.instance().log(this, getErrorLevel(), "TEST: Received hot resync request");
               hotResync = true;
             } else if (type == TYPE.REPLICA_FULL_RESYNC) {
-              LogManager.instance().log(this, Level.FINE, "TEST: Received full resync request");
+              LogManager.instance().log(this, getErrorLevel(), "TEST: Received full resync request");
               fullResync = true;
             }
           }
@@ -85,27 +83,28 @@ public class ReplicationServerReplicaRestartForceDbInstallIT extends Replication
       });
 
     if (server.getServerName().equals("ArcadeDB_0"))
-      server.registerTestEventListener(new TestCallback() {
+      server.registerTestEventListener(new ReplicationCallback() {
         @Override
         public void onEvent(final TYPE type, final Object object, final ArcadeDBServer server) {
+          if (!serversSynchronized)
+            return;
+
           // AS SOON AS SERVER 2 IS OFFLINE, A CLEAN OF REPLICATION LOG AND RESTART IS EXECUTED
           if ("ArcadeDB_2".equals(object) && type == TYPE.REPLICA_OFFLINE && firstTimeServerShutdown) {
-            LogManager.instance()
-                .log(this, Level.FINE, "TEST: Stopping Replica 2, removing latency, delete the replication log file and restart the server...");
+            LogManager.instance().log(this, Level.SEVERE,
+                "TEST: Stopping Replica 2, removing latency, delete the replication log file and restart the server...");
             slowDown = false;
             firstTimeServerShutdown = false;
 
             executeAsynchronously(() -> {
               getServer(2).stop();
-              GlobalConfiguration.HA_REPLICATION_QUEUE_SIZE.setValue(512);
-              try {
-                Thread.sleep(1000);
-              } catch (final InterruptedException e) {
-                LogManager.instance().log(this, Level.SEVERE, "TEST: ArcadeDB_0 HA event listener thread interrupted");
-                Thread.currentThread().interrupt();
-              }
+              GlobalConfiguration.HA_REPLICATION_QUEUE_SIZE.reset();
+
               Assertions.assertTrue(new File("./target/replication/replication_ArcadeDB_2.rlog.0").exists());
               new File("./target/replication/replication_ArcadeDB_2.rlog.0").delete();
+
+              LogManager.instance().log(this, Level.SEVERE, "TEST: Restarting Replica 2...");
+
               getServer(2).start();
               return null;
             });
